Get app details

Retrieve detailed information on an application in the steam store.

appdetails uses Steam's unofficial API.

appdetails_steamcmd uses the unofficial steamcmd API, which provides data from the Steam client network (powered by the steam Python library).

appdetails_steamspy uses the unofficial SteamSpy API, which provides custom estimates on owners, tags, genres and price dynamics. To get a full app list from SteamSpy, see steamspy.

appdetails(
  appid,
  client = NULL,
  filters = NULL,
  country_code = "US",
  language = "english"
)

appdetails_steamcmd(appid)

appdetails_steamspy(appid)

Arguments

appid

A single appID of an application in the Steam store.

client

Unknown.

filters

Keys to include in the output. Can be any key that is returned by this function. If basic, includes the following default keys: type, name, steam_appid, required_age, dlc, detailed_description, about_the_game, short_description, supported_languages, header_image, website, pc_requirements, mac_requirements, linux_requirements. filters does not accept any of these keys as input.

country_code

ISO 3166 country code representing the country from which to view the Steam store.

language

ISO 639-1 language code all tokenized strings should be returned in. Formally, a tokenized string is a string that is prefixed with a # in the VDF file. Not all tokenized strings have a translation for all languages. If no translation is available, defaults to English.

Value

appdetails

A nested list containing the following keys:

type: Type of application (game, dlc, demo, advertising, mod, or video)

name: Name of the application

steam_appid: appID of the application

required_age: Required age to view the application

controller_support: Whether the app has controller support (partial or full). May be omitted.

is_free: Whether the game is free to play

dlc: Array of appIDs of DLCs associated with this application.

detailed_description: A detailed description in HTML

about_the_game: "About the game" description in HTML

short_description: Short description in HTML, usually found in the application preview

fullgame: Name and appID of the full game. Sometimes found in demos or movies.

supported_languages: Languages that the application supports, in HTML.

header_image: URL to the header image.

capsule_image: URL to the capsule image.

capsule_imagev5: URL to a different capsule image.

website: URL to the application website.

pc_requirements: Minimum and recommended requirements for Windows, in HTML

mac_requirements: Minimum and recommended requirements for Mac OS, in HTML

linux_requirements: Minimum and recommended requirements for Linux, in HTML

developers: Names of the developers.

publishers: Names of the publishers.

demos: List of appIDs which are demos of this application.

price_overview: List containing the initial and final price as well as price discounts. Omitted for free to play games.

packages: List of packageIDs that include this game.

package_groups: Dataframe specifying purchase options.

platforms: Specifies whether the application is compatible with linux and mac.

metacritic: Metacritic score and URL.

categories: Store categories of the application. May be omitted.

genres: Genres of the application. May be omitted.

screenshots: A dataframe containing the paths to all showcase screenshots. Omitted if no screenshots exist.

movies: A dataframe containing the paths to all showcase movies. Omitted if no movies exist.

recommendations: A list containing the total number of recommendations.

recommendations: A list containing the total and highlighted achievements.

release_date: A list containing the release date and whether the application is coming soon.

support_info: A list containing support URL and email.

background: URL to the background image.

background_raw: URL to the raw background image.

content_descriptors: Content descriptor codes given to this application. Also includes a content descriptor description. A list of content descriptors is included in content_descriptors.

ratings: A list containing maturity ratings.

appdetails_steamspy

A dataframe containing information about name, developer, publisher, reviews, concurrent players, genres, tags and estimated owners.

appdetails_steamcmd

A nested list containing information listed here.

Note

appdetails is rate-limited at 200 requests per 5 minutes.

appdetails_steamcmd is not rate-limited.

appdetails_steamspy is rate-limited at 1 request per second.

Examples

if (FALSE) {
# returns the detailed description about team fortress
appdetails(440, filters = "detailed_description")

# returns the metacritic score and total number of reviews of Counter-Strike
appdetails(10, filters = c("recommendations", "metacritic"))

# returns steamspy data
appdetails_steamspy(440)

# returns steamcmd data
appdetails_steamcmd(440)
}