Events

Fetches data about news and events from the Steam News Hub.

query_events queries all events based on arbitrary filters. get_best_events fetches only the best events, whatever that means. get_event_details returns details about a number of events within a number of groups.

query_events(
  min_time = NULL,
  max_time = Sys.time(),
  ascending = FALSE,
  max_results = NULL,
  populate = FALSE,
  app_types = NULL,
  event_types = NULL,
  collection_id = NULL,
  sale_id = NULL
)

get_best_events(
  language = "english",
  include_steam_blog = TRUE,
  recency = 0,
  only_game_updates = FALSE
)

get_event_details(gid, clanid)

Arguments

min_time, max_time

Time range for which to include events.

ascending

Whether to sort event dates in an ascending order.

max_results

Maximum number of events to return. A minimum of 10 documents will always be returned. The maximum number of documents is restricted to 500.

populate

Number of events to populate with additional event data. Adds a a field events to the data that essentially contains the output of get_event_details for the specified number of events.

app_types

Vector of app types to be returned. Can include library, wishlist, following, recommended, steam, curator and featured.

event_types

Vector of event types. Can include a combination of numbers representing event types. That's all I know.

collection_id

Unknown.

sale_id

Unknown.

language

Language in which event data should be returned.

include_steam_blog

Whether events from the Steam blog should be included.

recency

Specifies the number of days that game must have been last played to be included in the query.

only_game_updates

Whether to filter out all events that are not game updates.

gid

Vector of event IDs for which to provide details.

clanid

List of group IDs for which the events in gid were created.

Value

query_events returns a list of dataframes. documents contains information on the event documents. apps contains information about the event apps. If authenticated, also shows information about last playtime if applicable. clans shows information about the group of each event. If populate is larger than 0, also shows an events field, that shows resolved information about each document including their text body.

get_best_events and get_event_details essentially return a list in the same format as the events field in query_events.

Examples

if (FALSE) { # \dontrun{
# set a time filter
query_events(
  min_time = as.POSIXct("2023-01-01"),
  max_time = as.POSIXct("2023-01-31")
)

# get event details about first 20 events
query_events(populate = 20)

# show only events of games in library
query_events(app_types = "library")
} # }