App list

Get a list of applications in the Steam store, either a complete list or the top games by concurrent players, all-time players, or top releases.

get_app_list()

get_games_by_ccu(
  language = "english",
  elanguage = NULL,
  country_code = "US",
  steam_realm = 1L,
  include = NULL,
  apply_user_filters = FALSE
)

get_most_played_games()

get_top_releases()

get_apps_in_genre(genre, language = "english", country_code = "US")

get_apps_in_category(category, language = "english", country_code = "US")

Arguments

language

ISO 639-1 language code all tokenized strings should be returned in. Formally, a tokenized string is a string that is prefixed with a # in the VDF file. Not all tokenized strings have a translation for all languages. If no translation is available, defaults to English.

elanguage

Numeric code representing the store language. A list of language codes and their corresponding languages is defined in ELanguage.

country_code

ISO 3166 country code representing the country from which to view the Steam store.

steam_realm

Number describing the Steam realm. A value of 1 indicates Steam Global, 2 indicates Steam China, and 0 indicates Unknown.

Value

get_app_list

A dataframe containing the appID and name.

get_games_by_ccu

A dataframe containing the appID, the rank by CCU and the total number and all-time record of concurrent players.

get_most_played_game

A dataframe containing the appID, the rank by total all-time players and players last week as well as the all-time record of concurrent players.

get_top_releases

A dataframe with three rows containing the top releases of the last three months. Each row contains the start of the month, the url path for the top releases and a vector top release appIDs. The URL path can be appended to the following URL:

https://store.steampowered.com/charts/topnewreleases/

See also

Examples

get_app_list()
#> # A tibble: 225,088 × 2
#>    appid name                 
#>    <int> <chr>                
#>  1     5 Dedicated Server     
#>  2     5 Dedicated Server     
#>  3     7 Steam Client         
#>  4     7 Steam Client         
#>  5     8 winui2               
#>  6     8 winui2               
#>  7    10 Counter-Strike       
#>  8    10 Counter-Strike       
#>  9    20 Team Fortress Classic
#> 10    20 Team Fortress Classic
#> # ℹ 225,078 more rows

# get most played games
get_games_by_ccu()
#> # A tibble: 100 × 4
#>     rank   appid concurrent_in_game peak_in_game
#>    <int>   <int>              <int>        <int>
#>  1     1     730             979102      1460550
#>  2     2     570             576248       600334
#>  3     3 2694490             441001       498966
#>  4     4 2767030             308322       460883
#>  5     5  578080             279189       747601
#>  6     6  252490             138269       144469
#>  7     7  271590             135117       178332
#>  8     8 2923300             116247       129339
#>  9     9 2507950              92733        98930
#> 10    10 1938090              83414       122668
#> # ℹ 90 more rows

# get most popular games of all time
get_most_played_games()
#> # A tibble: 100 × 4
#>     rank   appid last_week_rank peak_in_game
#>    <int>   <int>          <int>        <int>
#>  1     1     730              1      1460230
#>  2     2 2767030             -1       452007
#>  3     3     570              2       618358
#>  4     4  578080              3       750493
#>  5     5  431960              4       118350
#>  6     6 2694490             -1       498966
#>  7     7 1203220              5       298926
#>  8     8 1938090              6       137453
#>  9     9  271590              7       161127
#> 10    10 1172470              8       144354
#> # ℹ 90 more rows

# get the best releases
get_top_releases()
#> # A tibble: 3 × 4
#>   name                           start_of_month      url_path           item_ids
#>   <chr>                          <dttm>              <chr>              <list>  
#> 1 Top Releases of October 2024   2024-10-01 07:00:00 top_october_2024   <int>   
#> 2 Top Releases of September 2024 2024-09-01 07:00:00 top_september_2024 <int>   
#> 3 Top Releases of August 2024    2024-08-01 07:00:00 top_august_2024    <int>