Get reviews

Retrieve user reviews for a given Steam application.

get_app_reviews retrieves up to 100 reviews.

get_all_app_reviews has the ability to return much more reviews depending on input parameters.

get_app_reviews(
  appid,
  day_range = NULL,
  start_date = NULL,
  end_date = NULL,
  date_range_type = NULL,
  filter = "all",
  language = "all",
  review_type = "all",
  purchase_type = "steam",
  playtime = c(0, 0),
  filter_offtopic_activity = TRUE,
  num_per_page = 20,
  cursor = NULL
)

get_all_app_reviews(
  appid,
  day_range = NULL,
  start_date = NULL,
  end_date = NULL,
  date_range_type = NULL,
  filter = "recent",
  language = "all",
  review_type = "all",
  purchase_type = "steam",
  playtime = c(0, 0),
  filter_offtopic_activity = TRUE,
  max_tries = 10
)

get_review_histogram(appid, language = "english", review_score_preference = 2)

Arguments

appid

Application ID referencing a Steam application. There are generally two ways of retrieving an application ID. The first one is to inspect the URL of a Steam game. For example the application ID of Counter-Strike is 10: https://store.steampowered.com/app/10/CounterStrike/.

It is also possible to retrieve application IDs programmatically by exploiting Steam's search API. For details see search_apps.

day_range

Number of days from now to n days ago to look for reviews.

start_date

Date-time object describing the earliest time to look for reviews. Ignored if date_range_type = "all".

end_date

Date-time object describing the latest time to look for reviews. Ignored if date_range_type = "all".

date_range_type

Specifies what to do with reviews within the date range set by start_date and end_date. Options include include, exclude, and all. If all, ignores end_date and start_date.

filter

Specifies the sort order. Can be one of summary, recent, updated, and all. summary returns a summary of 10 most helpful comments. all sorts by helpfulness and applies moving windows in order to always return reviews. recent sorts by creation time and updated sorts by update time. For get_all_app_reviews, filter must be either updated or recent.

language

Which review language to include in the output. A full list of platform supported languages can be found in the Steamworks documentation. Defaults to "all", which removes the language filter.

review_type

Filters by the type of review. Can be either positive, negative, or all.

purchase_type

Filters by the type of purchase. Can be either steam, non_steam_purchase, or all.

playtime

Length-2 numeric vector specifying the minimum and maximum playtime in hours that the reviewer should have played the game. If 0, no required (minimum or maximum) play time is set.

filter_offtopic_activity

Whether to filter offtopic activity. This especially includes so-called "review bombs", i.e. many reviews written by users in a very short time to artificially alter the review score of a game. See the Steam blog entry for more information.

num_per_page

By default, Steam only shows 20 reviews per query. By setting num_per_page, this number can go up to a maximum of 100. For larger queries, consider iterating using get_all_app_reviews.

cursor

For each review page query, get_app_reviews returns a cursor ID. This ID can be used to paginate and iterate through many review pages at once. This argument is probably useless for most use cases and is extensively used for iteration purposes by get_all_app_reviews.

max_pages

Maximum number of pages until breaking the iteration and returning the result up to this point. You may set this to Inf to retrieve all reviews for a given query.

Value

A dataframe containing IDs, texts, and metadata of application reviews on Steam.

Examples

if (FALSE) { # \dontrun{
# reviews of counter-strike
get_app_reviews(10)

# get all negative bulgarian reviews
get_app_reviews(10, language = "bulgarian", review_type = "negative")

# get all recent reviews up to page 10
get_all_app_reviews(10, filter = "recent", max_tries = 10)
} # }