Discovery queue

Retrieve suggested and skipped apps on the discovery queue and its settings.

get_discovery_queue(
  queue_type = NULL,
  country_code = "US",
  rebuild_queue = FALSE,
  settings_changed = FALSE,
  os_win = FALSE,
  os_mac = FALSE,
  os_linux = FALSE,
  full_controller_support = FALSE,
  native_steam_controller = FALSE,
  include_coming_soon = FALSE,
  excluded_tagids = NULL,
  exclude_early_access = FALSE,
  exclude_videos = FALSE,
  exclude_software = FALSE,
  exclude_dlc = FALSE,
  exclude_soundtracks = FALSE,
  featured_tagids = NULL,
  rebuild_queue_if_stale = FALSE,
  ignore_user_preferences = FALSE,
  no_experimental_results = FALSE
)

get_discovery_settings(queue_type = NULL)

get_discovery_skipped(queue_type = NULL)

Arguments

queue_type

Integer indicating the queue type. Possible values can be retrieved using node_enum("StoreDiscoveryQueueType").

rebuild_queue

If TRUE, rebuilds the queue. Otherwise, takes the previously generated one. Rebuilding the queue can lead to error codes 2. In this case, retrying helps.

settings_changed

Whether to overwrite the settings in get_discovery_settings with the settings passed to get_discovery_queue.

os_win

If TRUE, only includes apps for Windows.

os_mac

If TRUE, only includes apps for Mac OS.

os_linux

If TRUE, only includes apps for Linux.

full_controller_support

If TRUE, only includes apps with full controller support.

native_steam_controller

If TRUE, only includes apps with native Steam controller support.

include_coming_soon

If TRUE, includes apps that are coming soon.

excluded_tagids

A vector of tagIDs to exclude.

exclude_early_access, exclude_videos, exclude_software, exclude_dlc, exclude_soundtracks

If TRUE, excludes early access apps, videos, software, DLCs, or soundtracks.

featured_tagids

A vector of tagIDs that must be present in the output appIDs.

rebuild_queue_if_stale

If TRUE, rebuilds the discovery queue if it is stale. It is a mystery what stale means.

ignore_user_preferences

If TRUE, ignores user preferences.

no_experimental_results

Unknown.

Value

get_discovery_queue

A vector of appIDs in the requested discovery queue.

get_discovery_settings

A list of current discovery queue settings.

get_discovery_skipped

A vector of appIDs skipped in the discovery queue.

Examples

# return discovery queue only containing DLCs on sale
get_discovery_queue(queue_type = 8) # 8 - DLCOnSale
#> Error: Session is not authenticated. You can login using the auth_credentials or auth_qr functions.

# return the best DLCs on sale in Germany
get_discovery_queue(queue_type = 8, country_code = "DE")
#> Error: Session is not authenticated. You can login using the auth_credentials or auth_qr functions.

if (FALSE) { # \dontrun{
# overwrite discovery queue settings
# NOTE: this modifies account settings and should be used with care
get_discovery_queue(settings_changed = TRUE, exclude_dlc = TRUE)
} # }