Get store items

Get information about items in the store. get_hardware_items more specifically requests special information about items that require physical shipping.

get_items(items, context = NULL, data_request = NULL)

get_hardware_items(items, context = NULL)

Arguments

items

A dataframe, list, or vector containing item IDs. If a vector, interprets it as appIDs. If a list or character vector is passed, each column can refer to a different type of item ID. Available types are appid, packageid, bundleid, tagid, creatorid, and hubcategoryid. In this way, multiple items with multiple ID types can be specified. For get_hardware_items, only vectors are allowed and they are always interpreted as packageIDs.

context

Object of class StoreBrowseContext that specifies the geographic context from which to access the store. Defaults to a global Steam realm (steam_realm = 1), English as the store language (language = "english") and the USA as the access country (country_code = "US").

data_request

Object of class StoreBrowseDataRequest that specifies the additional information to be included. Defaults to basic information.

Value

A dataframe containing the requested information on the input IDs.

Examples

# if just an ID is passed, it is assumed to be an appID
get_items(10)
#> # A tibble: 1 × 18
#>   item_type    id success visible name           store_url_path      appid  type
#>       <int> <int>   <int> <lgl>   <chr>          <chr>               <int> <int>
#> 1         0    10       1 TRUE    Counter-Strike app/10/CounterStri…    10     0
#> # ℹ 10 more variables: content_descriptorids <list>,
#> #   categories.supported_player_categoryids <list>,
#> #   categories.feature_categoryids <list>,
#> #   best_purchase_option.packageid <int>,
#> #   best_purchase_option.purchase_option_name <chr>,
#> #   best_purchase_option.final_price_in_cents <chr>,
#> #   best_purchase_option.formatted_final_price <chr>, …

# by passing a dataframe, you can pass multiple types of ID at a time
ids <- data.frame(
  appid = c(10, NA, NA),
  creatorid = c(NA, 10, NA),
  packageid = c(NA, NA, 354231)
)

# request basic item info
get_items(ids)
#> # A tibble: 3 × 22
#>   item_type     id success visible name               store_url_path appid  type
#>       <int>  <int>   <int> <lgl>   <chr>              <chr>          <int> <int>
#> 1         0     10       1 TRUE    Counter-Strike     app/10/Counte…    10     0
#> 2         5     10       1 TRUE    Seattle            curator/10        NA    NA
#> 3         1 354231       1 TRUE    Valve Index VR Kit sub/354231/       NA    10
#> # ℹ 14 more variables: content_descriptorids <list>, game_count <int>,
#> #   included_types <list>, included_appids <list>,
#> #   categories.supported_player_categoryids <list>,
#> #   categories.feature_categoryids <list>,
#> #   best_purchase_option.packageid <int>,
#> #   best_purchase_option.purchase_option_name <chr>,
#> #   best_purchase_option.final_price_in_cents <chr>, …

# request info in german language
get_items(ids, context = store_context(language = "german"))
#> Error in check_class(context, "StoreBrowseContext", null = TRUE): context must be of class StoreBrowseContext, not list

# request info for the swedish store
get_items(ids, context = store_context(country_code = "SE"))
#> Error in check_class(context, "StoreBrowseContext", null = TRUE): context must be of class StoreBrowseContext, not list

# request info on operating systems
get_items(ids, data_request = store_data_request(include = "platforms"))
#> Error in store_data_request(include = "platforms"): could not find function "store_data_request"

# request special info on hardware
get_items(ids, data_request = store_data_request(include = "all"))
#> Error in store_data_request(include = "all"): could not find function "store_data_request"
get_hardware_items(c(354231, 1628580))
#> Error: Steam Web API returned error code 8: InvalidParam