Get user info

Functions to retrieve information on a given user including friends, stats, achievements, groups.

get_player_summary(steamids)

get_player_bans(steamids)

get_owned_games(
  steamid,
  include_appinfo = FALSE,
  include_played_free_games = FALSE,
  appids_filter = NULL,
  include_free_sub = FALSE,
  skip_unvetted_games = FALSE,
  language = "english",
  include_extended_appinfo = FALSE
)

Arguments

steamid

Steam user ID as returned by get_steam_id.

relationship

Type of relationship to the user. One of all and friend.

Value

get_friend_list returns a dataframe on relationship status and time.

get_user_stats_for_game returns a list with names of the type of statistic, its value, and the type of statistic (achievement / stat).

get_user_group returns a character vector of group IDs.

get_player_summary returns a dataframe with one row for each user containing basic information on the user account.

get_player_bans returns a dataframe containing information on bans (e.g. VAC, community bans)

Examples

if (FALSE) { # \dontrun{
# get steam ID
steamid <- get_steam_id("gabelogannewell")

# get all friends
get_friend_list(steamid)

# user stats for team fortress
get_user_stats_for_game
} # }