Steam levels

Get steam levels of users, their percentiles and distributions.

get_steam_level retrieves the level of a user.

steam_level_percentile retrieves the percentile of a user level compared to all other user levels.

steam_level_distribution retrieves the user level percentile for a number of levels and shows their distribution.

get_steam_level(steamid)

steam_level_distribution(levels = 1:100)

get_level_percentile(level)

Arguments

steamid

SteamID of a user. The SteamID must be in a format that can be converted by convert_steamid. This includes vanity, Steam64, Steam2, and Steam3 IDs.

levels

A vector of user levels to show the distribution of.

level

User level to retrieve.

Value

get_steam_level and get_level_percentile return a length-1 numeric vector. steam_level_distribution returns a numeric vector of length length(levels).

Examples

if (FALSE) {
steamid <- "76561197984981409"

# get level of a user
get_steam_level("76561197984981409")

# get percentile of a user
get_level_percentile("76561197984981409")

# plot distribution
dist <- steam_level_distribution(1:20)
plot(dist)
}