Weekly top sellers

Retrieve the top 100 weekly top sellers of a given country.

get_weekly_top_sellers(
  country = NULL,
  language = "english",
  elanguage = NULL,
  country_code = "US",
  steam_realm = 1L,
  include = NULL,
  apply_user_filters = FALSE,
  start_date = NULL
)

Arguments

country

ISO-2 code of the country for which to return top sellers. A list of Steam countries can be retrieved with get_country_list.

start_date

Date-time object giving the earliest time for which to retrieve top sellers. Results will be returned for the time between the start date and the start date plus seven days.

page_start

On which part to start looking. Defaults to page 1.

page_count

Maximum pages to return. Defaults to one page.

Value

A dataframe containing information on weekly top sellers. Metadata is stored in attributes: timestamp holds the earliest time of the week, next_page_start specifies where the next page would start.

See also

Examples

# worldwide weekly top sellers
get_weekly_top_sellers()
#> Error in store_browse_context(language = language, elanguage = elanguage,     country_code = country_code, steam_realm = steam_realm): could not find function "store_browse_context"

# weeky top sellers of South Korea
get_weekly_top_sellers("KR")
#> Error in store_browse_context(language = language, elanguage = elanguage,     country_code = country_code, steam_realm = steam_realm): could not find function "store_browse_context"

# top sellers of last month
get_weekly_top_sellers(start_date = Sys.time() - 2629800)
#> Error in store_browse_context(language = language, elanguage = elanguage,     country_code = country_code, steam_realm = steam_realm): could not find function "store_browse_context"

# return additional info
get_weekly_top_sellers(include = "screenshots")
#> Error in store_browse_context(language = language, elanguage = elanguage,     country_code = country_code, steam_realm = steam_realm): could not find function "store_browse_context"