Assets

Query market assets and their prices. An asset is an item from the Steam economy that can be traded. For more information about assets, see the Steamworks documentation.

query_market_assets(
  appid,
  search_term = NULL,
  sort_column = NULL,
  sort_dir = NULL,
  search_description = NULL,
  paginate = TRUE
)

get_asset_prices(appid, currency = NULL, language = "english")

get_asset_info(
  appid,
  classids,
  class_count = 1L,
  language = "english",
  instanceids = NULL
)

get_price_history(appid, market_hash_name)

Arguments

appid

A single appID of an application in the Steam store.

search_term

Term to search for in the market item name.

sort_column

Column to sort the response by.

sort_dir

Direction of sorting if sort_column is not NULL. Can be asc or desc.

currency

ISO 4217 code to filter by currency. Defaults to all available currencies.

language

ISO 639-1 language code all tokenized strings should be returned in. Formally, a tokenized string is a string that is prefixed with a # in the VDF file. Not all tokenized strings have a translation for all languages. If no translation is available, defaults to English.

classids

A vector of classIDs of the assets to retrieve.

instanceids

A vector of instanceIDs corresponding to the

search_descriptions

Unknown.

Value

query_market_assetsA dataframe with each row representing one asset. The dataframe provides data on name, hash names, prices, the corresponding app, asset classIDs, and the output of get_asset_info.

get_asset_prices

A dataframe containg information on assert prices. name, date, class, and classid provide metadata about assets. Optionally, tags are provided in tags and tag_ids. class, tags, and tag_ids are nested as dataframes or lists as they can contain values of arbitrary length. prices.\* contain the current prices for the currencies specified in currency. If an asset is on sale, original_prices.\* show the original prices.

get_asset_info

A dataframe with each row representing asset info on each provided classID. \*_url columns show extensions to the URL http://cdn.steamcommunity.com/economy/image/. name, market_hash_name and market_name show various types of asset names. If an asset has received a fraud warning, the column fraudwarnings can contain their warning strings. The descriptions columns contain information about asset attributes, actions about action URLs, and tags about asset tags.

References

https://wiki.teamfortress.com/wiki/WebAPI/GetAssetPrices

https://wiki.teamfortress.com/wiki/WebAPI/GetAssetClassInfo

See also

Examples

# show asset prices for Team Fortress and all available currencies
get_asset_prices(440)
#> # A tibble: 1,203 × 80
#>    name  date       class classid tags  tag_ids prices.USD prices.GBP prices.EUR
#>    <chr> <date>     <lis> <chr>   <lis> <list>       <int>      <int>      <int>
#>  1 30186 2013-10-09 <df>  211447… <chr> <dbl>         1299       1039       1239
#>  2 341   2010-12-20 <df>  195151  <chr> <dbl>          499        399        475
#>  3 610   2011-09-07 <df>  9240401 <chr> <dbl>          749        599        715
#>  4 5873  2017-09-13 <df>  256945… <chr> <dbl>          249        199        235
#>  5 5082  2012-11-29 <df>  996660… <chr> <dbl>          499        399        475
#>  6 708   2012-01-26 <df>  329141… <chr> <dbl>          499        399        475
#>  7 405   2011-03-17 <df>  148     <chr> <int>           99         79         95
#>  8 30331 2013-12-19 <df>  260681… <chr> <dbl>          749        599        715
#>  9 993   2012-12-20 <df>  107343… <chr> <dbl>          749        599        715
#> 10 571   2011-10-28 <df>  168910… <chr> <dbl>          749        599        715
#> # ℹ 1,193 more rows
#> # ℹ 71 more variables: prices.RUB <int>, prices.BRL <int>, prices.JPY <int>,
#> #   prices.NOK <int>, prices.IDR <int>, prices.MYR <int>, prices.PHP <int>,
#> #   prices.SGD <int>, prices.THB <int>, prices.VND <int>, prices.KRW <int>,
#> #   prices.UAH <int>, prices.MXN <int>, prices.CAD <int>, prices.AUD <int>,
#> #   prices.NZD <int>, prices.PLN <int>, prices.CHF <int>, prices.CNY <int>,
#> #   prices.TWD <int>, prices.HKD <int>, prices.INR <int>, prices.AED <int>, …

# show the same assets, but return only EUR
get_asset_prices(440, currency = "EUR")
#> # A tibble: 1,203 × 8
#>    name  date       class  classid  tags  tag_ids prices.EUR original_prices.EUR
#>    <chr> <date>     <list> <chr>    <lis> <list>       <int>               <int>
#>  1 30186 2013-10-09 <df>   2114477… <chr> <dbl>         1239                  NA
#>  2 341   2010-12-20 <df>   195151   <chr> <dbl>          475                  NA
#>  3 610   2011-09-07 <df>   9240401  <chr> <dbl>          715                  NA
#>  4 5873  2017-09-13 <df>   2569452… <chr> <dbl>          235                  NA
#>  5 5082  2012-11-29 <df>   99666069 <chr> <dbl>          475                  NA
#>  6 708   2012-01-26 <df>   32914110 <chr> <dbl>          475                  NA
#>  7 405   2011-03-17 <df>   148      <chr> <int>           95                  NA
#>  8 30331 2013-12-19 <df>   2606815… <chr> <dbl>          715                  NA
#>  9 993   2012-12-20 <df>   1073435… <chr> <dbl>          715                  NA
#> 10 571   2011-10-28 <df>   16891096 <chr> <dbl>          715                  NA
#> # ℹ 1,193 more rows

# a list of valid currency codes can be retrieved using
# steamkit_enum
steamkit_enum("CurrencyCode", type = "SteamLanguage")
#> # A tibble: 47 × 2
#>    code  desc   
#>    <chr> <fct>  
#>  1 0     Invalid
#>  2 1     USD    
#>  3 2     GBP    
#>  4 3     EUR    
#>  5 4     CHF    
#>  6 5     RUB    
#>  7 6     PLN    
#>  8 7     BRL    
#>  9 8     JPY    
#> 10 9     NOK    
#> # ℹ 37 more rows

# get_asset_prices returns classIDs for each asset. These classIDs can
# be used to retrieve detailed information on an asset.
get_asset_class(730, c("5710093913", "5189384166"))
#> Error in get_asset_class(730, c("5710093913", "5189384166")): could not find function "get_asset_class"