API requests

Send HTTP requests to any of the available Steam APIs. As of yet, steamr supports the Web API, storefront API, SteamSpy and steamcmd.net. Generic requests to any other URL can also be made.

request_webapi(
  api,
  interface,
  method,
  version = "v1",
  params = list(),
  http_method = "GET",
  simplify = TRUE,
  paginate = NULL,
  format = c("json", "xml", "vdf"),
  serror = TRUE,
  dry = FALSE
)

request_storefront(
  api,
  interface,
  method,
  params = list(),
  params_as_query = TRUE,
  simplify = TRUE,
  paginate = NULL,
  rate = NULL,
  dry = FALSE
)

request_steamspy(params)

request_generic(
  url,
  params = NULL,
  method = "GET",
  format = "json",
  ...,
  headers = NULL,
  dry = FALSE
)

public_api()

partner_api()

store_api()

comm_api()

valve_api()

Arguments

api

URL to a supported Steam API. Can be public_api() (public Web API), partner_api() (private Web API), store_api() (store.steampowered), comm_api() (steamcommunity), or valve_api() (valvesoftware). Can also be any other host.

interface

Interface containing the API method.

method

Method that is to be requested.

version

Version string of the API method. Usually, this is v1, sometimes it is v2, and in rare occasions it can go as high as v3.

params

A list of parameters to send with the request. Note that input_json blobs need to be pre-formatted. Most (but not all) methods require an API key in the key parameter. Authenticated methods require an access_token in access_token.

http_method

HTTP method to be used. The Steam Web API and storefront API only take GET and POST.

simplify

Whether to simplify the output or leave it as a nested list.

paginate

If specified, tries to paginate through response pages. Can be one of cursor, offset, and input_json. cursor takes the next_cursor value of the response and re-inserts it in the cursor parameter of the next request (see iterate_with_cursor). offset takes increments the page parameter by 1 with each request (see iterate_with_offset). input_json re-creates the input_json blob with an incremented page parameter with each request.

format

Format of the response. One of json, xml, or vdf. vdf is Steam's KeyValues format, which can be parsed using parse_vdf.

serror

Whether to throw a formatted error when the response has the X-eresult header signifying an API error.

dry

If TRUE does not run the request, but returns the result of a dry run to a local server. See req_dry_run.

params_as_query

In most storefront endpoints, parameters are passed as a URL query (i.e. domain.org/path?name=param). However, some endpoints expect them to be passed as a URL path (i.e. domain.org/path/param1/param2). If FALSE, constructs parameters as a URL path, otherwise as a query.

url

URL to send a request to.

...

Further arguments passed to the parsing function defined by format, i.e. resp_body_json, resp_body_xml or resp_body_string.

headers

Key-value pairs of additional headers to append to the request.

Value

The formatted response of the request. If format = "json", formats the response as a list or dataframe (depending on the simplify argument). If format = "xml", returns an xml

object of the xml2 library. If format = "vdf", returns the unformatted string of a VDF document that can be parsed with parse_vdf.

Examples

if (FALSE) {
# simple Web API request
request_webapi(
  api = public_api(),
  interface = "ISteamApps",
  method = "GetAppList",
  version = "v2"
)

# paginate through query results
# QueryRewardItems uses cursors but other methods
# also use page parameters
request_webapi(
  api = public_api(),
  interface = "ILoyaltyRewardsService",
  method = "QueryRewardItems",
  params = list(appids = 440),
  paginate = "cursor"
)

# send a storefront request
request_storefront(
  api = store_api(),
  interface = "search",
  method = "suggest",
  params = list(term = "team fortress")
)

# request_steamspy is a low-level alternative to the
# steamspy function
request_steamspy(list(request = "tag", tag = "Early Access"))

# request_generic exists to enable non-standard API queries
# the following example is used internally in auth_* functions
request_generic(
  "https://login.steampowered.com/jwt/finalizelogin",
  method = "POST",
  params = list(
    nonce = "<refresh token>",
    sessionid = "<sessionid>",
    redir = "https://steamcommunity.com/login/home/?goto="
  )
)
}