Bundles and packages

Resolve details of bundles and packages.

resolve_bundles(bundleids, country_code = "US", language = "english")

resolve_packages(packageids, country_code = "US", language = "english")

Arguments

bundleids

A vector of up to 200 bundleIDs.

packageids

A vector of up to 200 packageIDs. For package_details, this vector can be of any length.

Value

Both resolve_bundles and resolve_packages return detailed information on a bundle or a package including prices, release dates, compatibility information, discounts, and localization.

Examples

# resolve Graveyard Keeper bundle
resolve_bundles(8077)
#> # A tibble: 1 × 23
#>   packageids appids bundleid name    header_image_url main_capsule initial_price
#>   <list>     <list>    <int> <chr>   <chr>            <chr>                <int>
#> 1 <int [6]>  <int>      8077 Gravey… https://shared.… https://sha…          6194
#> # ℹ 16 more variables: final_price <int>, formatted_orig_price <chr>,
#> #   formatted_final_price <chr>, discount_percent <int>,
#> #   bundle_base_discount <int>, available_windows <lgl>, available_mac <lgl>,
#> #   available_linux <lgl>, support_vrhmd <lgl>, support_vrhmd_only <lgl>,
#> #   creator_clan_ids <list>, localized_langs <list>, coming_soon <lgl>,
#> #   library_asset <chr>, no_main_cap <lgl>, deck_compatibility_category <int>

# resolve OST from the Graveyard Keeper bundle
resolve_packages(295588)
#> # A tibble: 1 × 23
#>   appids    packageid name                 formatted_orig_price orig_price_cents
#>   <list>        <int> <chr>                <chr>                           <int>
#> 1 <int [1]>    295588 Graveyard Keeper OST $5.99                             599
#> # ℹ 18 more variables: formatted_final_price <chr>, release_date <chr>,
#> #   available_windows <lgl>, available_mac <lgl>, available_linux <lgl>,
#> #   support_vrhmd <lgl>, support_vrhmd_only <lgl>, creator_clan_ids <list>,
#> #   localized_langs <list>, final_price_cents <int>, coming_soon <lgl>,
#> #   no_main_cap <lgl>, deck_compatibility_category <int>,
#> #   discount_percent <int>, discount_end_rtime <int>, header_image_url <chr>,
#> #   main_capsule <chr>, library_asset <chr>

# get more details on OST
package_details(295588)
#> Error in package_details(295588): could not find function "package_details"