Search apps

Simple Steam store queries that return the best matches to a given search query. Uses Steam's unofficial API.

Steam provides different endpoints that essentially do the same, but in a slightly different manner.

search_apps has less filtering capabilities, provides no context and returns a slim output.

suggest provides many filtering capabilities, provides context and returns a slightly more detailed output.

store_search has no filtering capabilities, provides context and returns a detailed output.

search_apps(term, links = FALSE)

suggest(
  term,
  country_code = "US",
  language = "english",
  steam_realm = 1L,
  require_type = NULL,
  use_store_query = TRUE,
  use_search_spellcheck = TRUE,
  search_creators_and_tags = TRUE,
  links = FALSE
)

store_search(term, country_code = "US", language = "english")

Arguments

term

Search term used to query the Steam store.

links

Whether to include links to icon and logo in the output. Defaults to FALSE to avoid littering the output.

require_type

Which application types should be returned? Can be one of bundle, game, demo, mod dlc, software, video, hardware, series, and music.

use_store_query

Unknown parameter.

use_search_spellcheck

Whether search terms should be spellchecked.

search_creators_and_tags

Whether to search for the search term in creator names and tags.

Value

A dataframe containing the appid and application name and (optionally) links to application icon and logo.

See also

Examples

if (FALSE) { # \dontrun{
# returns a single result
search_apps("team fortress")

# returns multiple results
search_apps("team")

# returns the first 10 results
suggest("team")

# returns the first 10 most fitting DLCs
suggest("team", require_type = "dlc")

# translates results to German and converts dollar to euro
suggest("team", language = "german", country_code = "DE")

# provides detailed data
store_search("team")
} # }