Support FAQ

Search support FAQ articles.

search_support(search_text, elanguage = NULL, paginate = TRUE, max_pages = Inf)

Arguments

search_text

Search text to match FAQ articles by.

paginate

If TRUE, paginates through the results. Some methods provide data access in digestible chunks and do not return all available information at once. The paginate automatically retrieves all pages until max_pages is reached.

max_pages

Maximum number of pages to paginate. Ignored if paginate is FALSE. Defaults to Inf such that all available data are paginated.

Examples

# get help articles on portal 2
search_support("portal 2")
#> # A tibble: 14 × 5
#>    articleid           name                      content clan_accountid url_code
#>    <chr>               <chr>                     <chr>            <int> <chr>   
#>  1 5718041075055894628 Portal 2 Perpetual Testi… "Are y…              0 4F5A-90…
#>  2 1964164149307655285 Achievements Disappear i… "When …              0 1B42-1C…
#>  3 4198363033106924797 Crosshair disappears whe… "When …              0 3A43-94…
#>  4 3363261774082873484 NVIDIA Shield             "Games…              0 2EAC-B4…
#>  5 381734390587148756  Steam Families User Guid… "Steam…              0 054C-31…
#>  6 4448975342267632853 Unsupported Video Cards … "Steam…              0 3DBD-EF…
#>  7 1166099635598618949 Failed to create D3D Dev… "D3D e…              0 102E-D1…
#>  8 8978166077040914691 Windows Mixed Reality Se… "https…              0 7C98-DA…
#>  9 7393273865525372396 Troubleshooting Network … "Troub…              0 669A-2F…
#> 10 6057563872459458384 DirectX and Pixel Shader… "Pixel…              0 5410-CA…
#> 11 3839505547814736911 Steam Deck - Network + W… "Gener…              0 3548-A9…
#> 12 8201958456624065082 Limited User Accounts     "Steam…              0 71D3-35…
#> 13 6558267004196197409 Troubleshooting Game Per… "How c…              0 5B03-A5…
#> 14 4484950172723160530 Steam Link - Suggested N… "https…              0 3E3D-BE…

# get help articles on purchases
search_support("purchase")
#> # A tibble: 103 × 5
#>    articleid           name                      content clan_accountid url_code
#>    <chr>               <chr>                     <chr>            <int> <chr>   
#>  1 8294526361226486046 Purchasing Issues         "Gener…              0 731C-13…
#>  2 5481077739806853445 Steam Sales - Purchasing… "Commo…              0 4C10-B3…
#>  3 2796289223589124788 Retrying a failed purcha… "My pu…              0 26CE-6A…
#>  4 4926790029447666731 Your Purchase could not … "When …              0 445F-79…
#>  5 8604539892456401617 Steam Is Asking Me To Pu… "I bou…              0 7769-77…
#>  6 2059425740868674023 An unexpected error has … "What …              0 1C94-8C…
#>  7 3171155790013657367 Steam Gifts               "What …              0 2C02-35…
#>  8 4656875171981542682 Providing Proof of Owner… "If th…              0 40A0-8B…
#>  9 8201958456624065082 Limited User Accounts     "Steam…              0 71D3-35…
#> 10 8710933858579164505 Steam Wallet              "Steam…              0 78E3-74…
#> # ℹ 93 more rows

# get localized articles in french
search_support("purchase", elanguage = 2)
#> Iterating ■■■■■■■■■■■                       33% | ETA:  2s
#> Iterating ■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■  100% | ETA:  0s
#> # A tibble: 105 × 6
#>    articleid           name      content clan_accountid url_code localized_names
#>    <chr>               <chr>     <chr>            <int> <chr>    <list>         
#>  1 8294526361226486046 Problème… "Dépan…              0 731C-13… <chr [1]>      
#>  2 5481077739806853445 Promotio… "Quest…              0 4C10-B3… <chr [1]>      
#>  3 2796289223589124788 Effectue… "Mon a…              0 26CE-6A… <chr [1]>      
#>  4 4926790029447666731 Votre tr… "Quand…              0 445F-79… <chr [1]>      
#>  5 8604539892456401617 Steam me… "J'ai …              0 7769-77… <chr [1]>      
#>  6 2059425740868674023 Une erre… "Que f…              0 1C94-8C… <chr [1]>      
#>  7 3171155790013657367 Cadeaux … "Qu'es…              0 2C02-35… <chr [1]>      
#>  8 4656875171981542682 Fournir … "Si le…              0 40A0-8B… <chr [1]>      
#>  9 8201958456624065082 Comptes … "Steam…              0 71D3-35… <chr [1]>      
#> 10 8710933858579164505 Portemon… "Porte…              0 78E3-74… <chr [1]>      
#> # ℹ 95 more rows