Get store items

Get information about items in the store

get_items(
  items,
  language = "english",
  elanguage = NULL,
  country_code = "US",
  steam_realm = 1L,
  include = NULL,
  apply_user_filters = FALSE
)

Arguments

items

Store item or list of store items as returned by store_item.

language

ISO639-1 language code all tokenized strings should be returned in. Formally, a tokenized string is a string that is prefixed with a # in the VDF file. Not all tokenized strings have a translation for all languages. If no translation is available, defaults to English.

elanguage

Language identifier of the returned information.

country_code

ISO-2 country code as returned by get_countries. The Steam store differs depending on the country it is accessed from.

steam_realm

Number describing the Steam realm. A value of 1 indicates Steam Global, 2 indicates Steam China, and 0 indicates Unknown.

include

List of extra information to include. Can be one or several of the following: release, platforms, all_purchase_options, screenshots, trailers, ratings, tag_count, reviews, basic_info, supported_languages, full_description, included_items and assets_without_overrides. If all, returns all information.

apply_user_filters

Unknown.

Value

A dataframe containing the requested information on the input IDs.

See also

Examples

if (FALSE) { # \dontrun{
ids <- list(
  store_item(268770),
  store_item(1182620),
  store_item(creatorid = 10)
)

# request basic item info
get_items(ids)

# request info in german language
get_items(ids, language = "german")

# request info for the swedish store
get_items(ids, country_code = "SE")

# request info on operating systems
get_items(ids, include = "platforms")

# request all info
get_items(ids, include = "all")
} # }